# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  main.CustomTrack:
    properties:
      chineseName:
        type: string
      courseId:
        type: string
      createdAt:
        type: string
      custom:
        type: boolean
      duration:
        type: number
      fileSize:
        type: integer
      format:
        type: string
      id:
        type: string
      sortOrder:
        type: integer
      title:
        type: string
      unitId:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  main.ErrorResponse:
    description: API错误响应
    properties:
      error:
        description: 错误信息
        example: 未授权访问
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 提供语言学习音频播放器的后端API服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 语言学习音频播放器API
  version: "1.0"
paths:
...
  /api/courses/{courseId}/units/{unitId}/custom-tracks:
    get:
      consumes:
      - application/json
      description: 获取用户在指定课程和单元下的自定义音轨
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 自定义音轨列表
          schema:
            items:
              $ref: '#/definitions/main.CustomTrack'
            type: array
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 获取用户自定义音轨
      tags:
      - custom-tracks
    post:
      consumes:
      - multipart/form-data
      description: 上传并添加新的自定义音轨
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      - description: 音频文件
        in: formData
        name: audioFile
        required: true
        type: file
      - description: 音轨标题
        in: formData
        name: title
        required: true
        type: string
      - description: 中文标题
        in: formData
        name: chineseName
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "413":
          description: 文件过大
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 添加自定义音轨
      tags:
      - custom-tracks
  /api/courses/{courseId}/units/{unitId}/custom-tracks/{trackId}:
    delete:
      consumes:
      - application/json
      description: 删除用户创建的自定义音轨
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无权限操作
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音轨不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 删除自定义音轨
      tags:
      - custom-tracks
    patch:
      consumes:
      - application/json
      description: 更新自定义音轨的标题和中文名称
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      - description: 更新信息
        in: body
        name: track
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无权限操作
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音轨不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 更新自定义音轨
      tags:
      - custom-tracks
  /api/courses/{courseId}/units/{unitId}/custom-tracks/{trackId}/reorder:
    post:
      consumes:
      - application/json
      description: 调整自定义音轨的排序顺序
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      - description: 排序信息
        in: body
        name: order
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 排序成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无权限操作
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音轨不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 重排序自定义音轨
      tags:
      - custom-tracks
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"